requirements.txt
asgiref==3.4.1
autopep8==1.6.0
Django==3.2.9
djangorestframework==3.12.4
drf-nested-routers==0.93.4
pycodestyle==2.8.0
Pygments==2.10.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
typing-extensions==4.0.0
